% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AprioriTBDR.R
\name{AprioriTBDR}
\alias{AprioriTBDR}
\title{Apriori: Uma regra de associação aplicado á base em TBDR}
\usage{
AprioriTBDR(dataFrameTB, seedValue)
}
\arguments{
\item{\code{dataFrameTB}}{e \code{seedValue}}
}
\value{
Retornar um dataframe \code{dataFrameTB} de regras (Apriori aplicado)
}
\description{
Função aplica Apriori na base \code{dataFrameTB}
e setar o Seed \code{seedValue} para reproduzir a partição.
}
