% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criarTabelaAmostragem.R
\name{criarTabelaAmostragem}
\alias{criarTabelaAmostragem}
\title{TBDR data preparation}
\usage{
criarTabelaAmostragem(dataFrameTB, seedValue, sampleSize)
}
\arguments{
\item{\code{dataFrameTB}}{\code{sampleSize} e \code{seedValue}}
}
\value{
Retornar um dataframe de sample \code{dataFrameTB}
}
\description{
função para criar samples de uma base \code{dataFrameTB}
de tamanho \code{sampleSize} de \code{0.0 a 1.0}
e setar o Seed \code{seedValue} para reproduzir a partição
}
