% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutateFilesTBDR.R
\name{mutateFilesTBDR}
\alias{mutateFilesTBDR}
\title{Coincedente Mutation: Separe os itens nas regras para gerar uma matriz}
\usage{
mutateFilesTBDR(dataFrameTB)
}
\arguments{
\item{\code{ruleDataFrameTB}}{}
}
\value{
Retornar uma matriz em dataframe \code{matrizCoincedente}.
}
\description{
Função recebe um dataframe de coincedente \code{ruleDataFrameTB},
separe os itens nas regras do coincedente e gerar uma matriz de tal forma que
cada item virar uma coluna da matriz.
}
