% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegraValidaPreparacao.R
\name{prepararRegrasValidas}
\alias{prepararRegrasValidas}
\title{Regras Validas: Prepare as regras e separe por metricas e desfechos}
\usage{
prepararRegrasValidas(ruleDataFrameTB)
}
\arguments{
\item{\code{ruleDataFrameTB}}{}
}
\value{
Retornar um dataframe \code{data_rules} de regras, metricas e desfechos.
}
\description{
Função recebe um conjunto de regras \code{ruleDataFrameTB} logo
apos a aplicação do apriori, eliminação de '{' e '}', criar as colunas na ordem de
("Regras", "count", "lift","confidence", "support","Desfechos")
}
